hdfs.py refactoring: Move out target
